% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_local_prev.R
\name{calc_distance}
\alias{calc_distance}
\title{This function computes the distance of one ID versus the rest and obtains the n closest ones.}
\usage{
calc_distance(value, positions, test_result, thr_dist)
}
\arguments{
\item{value}{index detected by epifriends.}

\item{positions}{data.frame with the positions of parameters we want to query with shape (n,2) where n is the number of positions.}

\item{test_result}{data.frame with the test results (0 or 1).}

\item{thr_dist}{Maximum distance threshold.}
}
\value{
Prevalence based on n closest elements cross a given ID
}
\description{
This function computes the distance of one ID versus the rest and obtains the n closest ones.
}
\examples{
library(data.table)
# Creation of x vector of longitude coordinates, y vector of latitude coordinates and finaly merge them on a position data frame.
x <- c(1,2,3,4,7.5,8,8.5,9,10,13,13.1,13.2,13.3,14,15,30)
y <- c(1,2,3,4,7.5,8,8.5,9,10,13,13.1,13.2,13.3,14,15,30)
pos <- data.table("x" = x,"y" = y)
value <- c(1)

# Creation of test data frame with 0 for negative cases and 1 for positive clases for each position.
test <- data.frame(c(0,1,1,0,1,0,1,1,0,0,0,0,1,0,1,1))

# Creation of catalogue for this positions, linking distance 2 and default values.
pos_rate <- calc_distance(value, pos, test, 0.2)

}
\author{
Eric Matamoros.
}
