% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue.R
\name{catalogue}
\alias{catalogue}
\title{This method runs the DBSCAN algorithm (if cluster_id is NULL) and obtains the mean positivity rate (PR) of each cluster extended with the non-infected cases closer than the link_d.}
\usage{
catalogue(
  positions,
  test_result,
  link_d,
  cluster_id = NULL,
  min_neighbours = 2,
  max_p = 1,
  min_pos = 2,
  min_total = 2,
  min_pr = 0,
  keep_null_tests = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{positions}{data.frame with the positions of parameters we want to query with shape (n,2) where n is the number of positions.}

\item{test_result}{data.frame with the test results (0 or 1).}

\item{link_d}{The linking distance to connect cases. Should be in the same scale as the positions.}

\item{cluster_id}{Numeric vector with the cluster IDs of each position, with 0 for those without a cluster. Give NULL if cluster_id must be calculated.}

\item{min_neighbours}{Minium number of neighbours in the radius < link_d needed to link cases as friends.}

\item{max_p}{Maximum value of the p-value to consider the cluster detection.}

\item{min_pos}{Threshold of minimum number of positive cases in clusters applied.}

\item{min_total}{Threshold of minimum number of cases in clusters applied.}

\item{min_pr}{Threshold of minimum positivity rate in clusters applied.}

\item{keep_null_tests}{Whether to remove or not missings. If numeric, provide value to impute.}

\item{verbose}{If TRUE, print information of the process; else, do not print.}
}
\value{
List with the next objects:
cluster_id: numeric vector
  Vector of the cluster IDs of each position, with 0 for those
  without a cluster.
mean_pr_cluster: numeric vector
  Mean PR corresponding to cluster_id.
pval_cluster: numeric vector
  P-value corresponding to cluster_id.
epifriends_catalogue: List
  Catalogue of the epifriends clusters and their main characteristics.
}
\description{
This method runs the DBSCAN algorithm (if cluster_id is NULL) and obtains the mean positivity rate (PR) of each cluster extended with the non-infected cases closer than the link_d.
}
\details{
The epifriends package uses the RANN package which can gives the exact nearest neighbours using the friends of friends algorithm. For more information on the RANN library please visit https://cran.r-project.org/web/packages/RANN/RANN.pdf
#'
}
\examples{
# Required packages
if(!require("RANN")) install.packages("RANN")
library("RANN")

# Creation of x vector of longitude coordinates, y vector of latitude coordinates and finaly merge them on a position data frame.
x <- c(1,2,3,4,7.5,8,8.5,9,10,13,13.1,13.2,13.3,14,15,30)
y <- c(1,2,3,4,7.5,8,8.5,9,10,13,13.1,13.2,13.3,14,15,30)
pos <- data.frame(x,y)

# Creation of test data frame with 0 for negative cases and 1 for positive clases for each position.
test <- data.frame(c(0,1,1,0,1,0,1,1,0,0,0,0,1,0,1,1))

# Creation of catalogue for this positions, linking distance 2 and default values.
cat <- catalogue(pos, test, 2)
}
\author{
Mikel Majewski Etxeberria based on earlier python code by Arnau Pujol.
}
