% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datarand}
\alias{datarand}
\title{Dataset with random distribution of disease}
\format{
## `datarand`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{X, id}{Index or reference}
  \item{x, y}{Coordinates}
  \item{test}{Positive or negative tests}
  \item{case_count}{Number of tests with the given id}
  \item{dates}{Date of the test. Fixed dates}
  \item{geometry}{Geometrical way to express x & y coordinates}
}
}
\source{
Mock up dataset used for example used in epifriends_on_different_distributions.Rmd
}
\usage{
datarand
}
\description{
Random distribution of positive cases
}
\keyword{datasets}
