% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datarand_temp}
\alias{datarand_temp}
\title{Dataset with random distribution of positive disease cases distributed against different
timeframes.}
\format{
## `datarand_temp`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{X, id}{Index or reference}
  \item{x, y}{Coordinates}
  \item{test}{Positive or negative tests}
  \item{case_count}{Number of tests with the given id}
  \item{date}{Date of the test. There are different dates}
  \item{geometry}{Geometrical way to express x & y coordinates}
}
}
\source{
Mock up dataset used for example used in temporal_analysis.Rmd
}
\usage{
datarand_temp
}
\description{
Different clusters of positive cases distributed along negative ones with different
temporalities to try EpiFRIenDs on different timestamps.
}
\keyword{datasets}
