% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_catalogue.R
\name{distance}
\alias{distance}
\title{This method calculates the Euclidean distance between two positions.}
\usage{
distance(pos_a, pos_b)
}
\arguments{
\item{pos_a}{First position.}

\item{pos_b:}{Second position.}
}
\value{
Distance between positions.
}
\description{
This method calculates the Euclidean distance between two positions.
}
\examples{
# Creation of x vector of longitude coordinates, y vector of latitude coordinates and 
# finally merge them on a position data frame.
pos_a <- c(1,2)
pos_b <- c(4,7)

# Compute distances between two positions
x <- distance(pos_a, pos_b)

}
\author{
Eric Matamoros Morales based on earlier python code by Arnau Pujol.
}
