% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan.R
\name{find_indeces}
\alias{find_indeces}
\title{This method returns the indeces of all the friends of each position from positions given a KDTree.}
\usage{
find_indeces(positions, link_d, positions_eval)
}
\arguments{
\item{positions}{Data.frame with x & y coordinates}

\item{link_d:}{The linking distance to connect cases. Should be in the same scale as the positions.}

\item{positions_eval:}{Data.frame from which each x & y coordinate of positions wants to be evaluated}
}
\value{
List with an array of the indeces of the friends of each position.
}
\description{
This method returns the indeces of all the friends of each position from positions given a KDTree.
}
\details{
The epifriends package uses the RANN package which gives us the exact nearest neighbours using the friends of friends algorithm. For more information on the RANN library please visit https://cran.r-project.org/web/packages/RANN/RANN.pdf
}
\examples{
# Required packages
if(!require("RANN")) install.packages("RANN")
library("RANN")

# Creation of x vector of longitude coordinates, y vector of latitude coordinates and finaly merge them on a position data frame.
x <- c(1,2,3,4,7.5,8,8.5,9,10,13,13.1,13.2,13.3,14,15,30)
y <- c(1,2,3,4,7.5,8,8.5,9,10,13,13.1,13.2,13.3,14,15,30)
pos <- data.frame(x,y)

# Computation of clusters of hotspots for positions with dbscan algorithm using linking distance 2 and minimum 3 neighbours.
indeces <- find_indeces(pos, 0.1 ,pos)

}
\author{
Mikel Majewski Etxeberria based on earlier python code by Arnau Pujol.
}
