% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_catalogue.R
\name{get_label_list}
\alias{get_label_list}
\title{This method gives the unique values of a variable in a list of lists(temporal catalogues with temporal_id in our case).}
\usage{
get_label_list(df_list, label = "tempID")
}
\arguments{
\item{df_list}{List of lists(temporal catalogues with temporal_id in our case).}

\item{label}{Name of variable to select.}
}
\value{
List of unique values of the selected variable over all lists (temporal catalogues with temporal_id in our case) from the list.
}
\description{
This method gives the unique values of a variable in a list of lists(temporal catalogues with temporal_id in our case).
}
\author{
Eric Matamoros Morales based on earlier python code by Arnau Pujol.
}
