% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automate_link_d.R
\name{get_min_distances}
\alias{get_min_distances}
\title{This method optimizes the linking distance based on the distribution of the minimum distances that satisfy a given
minimum neighbor restriction.}
\usage{
get_min_distances(position, positive, min_neighbors)
}
\arguments{
\item{positive}{Array with location of positive tests.}

\item{positions}{data.frame with the positions of parameters we want to query with shape (n,2) where n is the number of positions.}

\item{min_neighbours}{Minium number of neighbours in the radius < link_d needed to link cases as friends.}
}
\value{
Minimum distance for each positive test where a given contrain of min_neighbors is satisfied.
}
\description{
This method optimizes the linking distance based on the distribution of the minimum distances that satisfy a given
minimum neighbor restriction.
}
\examples{
# Creation of x vector of longitude coordinates, y vector of latitude coordinates and finaly merge them on a position data frame.
x <- c(1,2,3,4,7.5,8,8.5,9,10,13,13.1,13.2,13.3,14,15,30)
y <- c(1,2,3,4,7.5,8,8.5,9,10,13,13.1,13.2,13.3,14,15,30)
test <-c(0,1,1,0,1,0,1,1,0,0,0,0,1,0,1,1)

df <- datatable(x = x, y = y, test = test)
position <- df[,.(x,y)]
positive = (df$test == 1)
min_neighbors <- 2

# Get minimum distances based on neighbors
min_distances <- get_min_distances(position, positive, min_neighbors)
}
\author{
Eric Matamoros.
}
