% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_coords.R
\name{latlon2xy}
\alias{latlon2xy}
\title{This method transforms the latitude and longitud coordinates from to cartesian coordinates in meters.}
\usage{
latlon2xy(lon, lat, to_epsg = NULL, verbose = TRUE)
}
\arguments{
\item{lon}{list of longitude positions.}

\item{lat}{List of latitude positions}

\item{to_epsg}{EPSG number for the projection to use}

\item{verbose}{It specifies if information on the process is printed}
}
\value{
List of coordinates x & y converted to cartesian coordinates
}
\description{
This method transforms the latitude and longitud coordinates from to cartesian coordinates in meters.
}
\examples{
# Required packages
if(!require("data.table")) install.packages("data.table")
library("data.table")

# Creation of x vector of longitude coordinates, y vector of latitude coordinates and finaly merge them on a position data frame.
x <- c(1,2,3,4,7.5,8,8.5,9,10,13,13.1,13.2,13.3,14,15,30)
y <- c(1,2,3,4,7.5,8,8.5,9,10,13,13.1,13.2,13.3,14,15,30)


# Convert to coordinates in meters
coords <- latlon2xy(x, y, 32733)

}
\author{
Eric Matamoros Morales based on earlier python code by Arnau Pujol.
}
