% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{lifetime_timeline}
\alias{lifetime_timeline}
\title{This method visualises the size evolution of the clusters in time and colour-codes 
them with their total lifetime.}
\usage{
lifetime_timeline(list_catalogues, mean_dates, time_steps)
}
\arguments{
\item{list_catalogues}{List of EpiFRIenDs catalogues, where each element contains 
the catalogue in each time step.}

\item{mean_dates}{List of dates corresponding to the mean time in each time window}

\item{time_steps}{Number of days that the time window is shifted in each time step}
}
\value{
ggplot object with the size evolution and lifetimes of clusters
}
\description{
This method visualises the size evolution of the clusters in time and colour-codes 
them with their total lifetime.
}
\examples{
# Creation of x vector of longitude coordinates, y vector of latitude coordinates 
# and finally merge them on a position data frame.
x <- c(1,2,3,4,7.5,8,8.5,9,10,13,13.1,13.2,13.3,14,15,30)
y <- c(1,2,3,4,7.5,8,8.5,9,10,13,13.1,13.2,13.3,14,15,30)
pos <- data.frame(x,y)
time_steps = 305

# Creation of test data frame with 0 for negative cases and 1 for positive 
# classes for each position.
test <- c(0,1,1,0,1,0,1,1,0,0,0,0,1,0,1,1)

#Creation of chron dates vector in format 'y-m-d h:m:s'.
dates <- c("2020-11-03 05:33:07","2021-05-19 10:29:59","2021-02-09 14:53:20",
"2021-11-21 02:35:38","2020-11-19 05:57:24", "2021-06-09 07:50:30",
"2021-09-18 05:53:26","2020-03-19 17:16:56","2021-06-08 12:40:46","2020-06-26 05:01:31",
"2020-10-15 04:40:27","2021-05-28 15:23:23","2020-05-01 02:56:54","2020-08-19 22:45:35",
"2021-10-23 18:56:35", "2020-10-19 00:01:25")

# Creation of temporal catalogue for this data.
tcat <- tcat <- temporal_catalogue(pos, test, dates ,link_d = 2, time_width = 305, 
time_steps = time_steps, linking_time = 3, linking_dist = 2)

lifetime_timeline(tcat$temporal_catalogues, tcat$mean_dates, time_steps)

}
\author{
Eric Matamoros Morales based on earlier python code by Arnau Pujol.
}
