% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mock_data_linkneigh}
\alias{mock_data_linkneigh}
\title{Dataset containing clusters of population with randomly distributed positive counts.}
\format{
## `mock_data_linkneigh`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{X, id}{Index or reference}
  \item{x, y}{Coordinates}
  \item{test}{Positive or negative tests}
  \item{case_count}{Number of tests with the given id}
  \item{date}{Date of the test. Fixed dates}
  \item{geometry}{Geometrical way to express x & y coordinates}
}
}
\source{
Mock up dataset used for example used in link_neighbours.Rmd
}
\usage{
mock_data_linkneigh
}
\description{
Some of the generated clusters contain a higher prevalence of the disease and others 
have less prevalence. The aim of this dataset is to mimic the differences between urban
& rural regions in real data.
}
\keyword{datasets}
