% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mock_data_local}
\alias{mock_data_local}
\title{Dataset with different local-fluctuations on the prevalence of the disease}
\format{
## `mock_data_local`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{X, id}{Index or reference}
  \item{x, y}{Coordinates}
  \item{test}{Positive or negative tests}
  \item{case_count}{Number of tests with the given id}
  \item{date}{Date of the test. Fixed dates}
  \item{geometry}{Geometrical way to express x & y coordinates}
}
}
\source{
Mock up dataset used for example used in local_prevalence.Rmd
}
\usage{
mock_data_local
}
\description{
The twho hemispheres along the Y-axis have clusters made with different fixed 
prevalences. Additionally, there is a new cluster in each hemisphere with the opposite
prevalence with the aim to add local deviations.
}
\keyword{datasets}
