% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mock_false_det}
\alias{mock_false_det}
\title{Dataset used for the example of False Detections}
\format{
## `mock_false_det`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{X, id}{Index or reference}
  \item{x, y}{Coordinates}
  \item{test}{Positive or negative tests}
  \item{case_count}{Number of tests with the given id}
  \item{date}{Date of the test. Fixed dates}
  \item{geometry}{Geometrical way to express x & y coordinates}
}
}
\source{
Mock up dataset used for example used in ldetect_false_positives.Rmd
}
\usage{
mock_false_det
}
\description{
This dataset contains differently generated clusters with random distribution of
positive cases & is being used for the False Detection example.
}
\keyword{datasets}
