% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot the coordinates with different colors for each of the clusters identified by the EpiFRIenDs.}
\usage{
plot_clusters(coordinates, catalogue, only_significant = FALSE)
}
\arguments{
\item{coordinates}{data frame with the values of the coordinates.}

\item{catalogue}{List of EpiFRIenDs catalogues.}

\item{only_significant}{If True, color only the clusters identified as significant by the algorithm.}
}
\value{
ggplot object with the clusters identified by the EpiFRIenDs algorithm.
}
\description{
Plot the coordinates with different colors for each of the clusters identified by the EpiFRIenDs.
}
\examples{
# Creation of x vector of longitude coordinates, y vector of latitude coordinates 
# and finally merge them on a position data frame.
x <- c(1,2,3,4,7.5,8,8.5,9,10,13,13.1,13.2,13.3,14,15,30)
y <- c(1,2,3,4,7.5,8,8.5,9,10,13,13.1,13.2,13.3,14,15,30)
pos <- data.frame(x,y)

# Creation of test data frame with 0 for negative cases and 1 for positive clases for each position.
test <- c(0,1,1,0,1,0,1,1,0,0,0,0,1,0,1,1)

# Creation of catalogue for this positions, linking distance 2 and default values.
cat <- catalogue(pos, test, 2)

plot_clusters(pos, cat)
}
\author{
Eric Matamoros Morales based on earlier python code by Arnau Pujol.
}
