% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automate_link_d.R
\name{quantile_estimation}
\alias{quantile_estimation}
\title{Obtain desired quantiles that will be used to evaluate distances.}
\usage{
quantile_estimation(quantile_est, verbose)
}
\arguments{
\item{quantile_est}{If 'fixed', quantiles used will be from 0 to 1 with a step 
of 0.125. If 'random', a random estimation of the step will be done.}

\item{verbose:}{If TRUE, print information of the process; else, do not print.}
}
\value{
Vector with quantiles
}
\description{
Obtain desired quantiles that will be used to evaluate distances.
}
\examples{
quantiles <- quantile_estimation('fixed', FALSE)
}
\author{
Eric Matamoros.
}
