% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scatter_pval}
\alias{scatter_pval}
\title{This method shows a scatter plot of the data distribution, showing in colour the positive cases that belong to foci (with the colour 
representing their p-value) and in grey the rest of the data.}
\usage{
scatter_pval(
  coordinates,
  id_data,
  positive,
  prevalence,
  epi_catalogue,
  use_geom_map = FALSE,
  lon_min = NULL,
  lon_max = NULL,
  lat_min = NULL,
  lat_max = NULL,
  xlims = NULL,
  ylims = NULL,
  title = NULL,
  region = "\\"Mozambique",
  thr_amplify = 0.05
)
}
\arguments{
\item{coordinates}{data frame with the values of the coordinates}

\item{id_data}{data frame with the cluster ID associated to each positive case, o for no associated cluster}

\item{positive}{Boolean vector that indicates if the case is infected.}

\item{epi_catalogue}{List of the EpiFRIenDs catalogue}

\item{use_geom_map:}{If True, use a geo-map to plot.}

\item{xlims:}{Vector with the limits of the X axis.}

\item{ylims:}{Vector with the limits of the Y axis}

\item{title:}{Title of the plot. If NULL, it defaults to 'P-value of hotspots'}
}
\value{
Scatter plot of the data distribution, showing in colour the positive cases that belong to foci (with the colour 
representing their p-value) and in grey the rest of the data.
}
\description{
This method shows a scatter plot of the data distribution, showing in colour the positive cases that belong to foci (with the colour 
representing their p-value) and in grey the rest of the data.
}
\examples{
# Required packages
if(!require("RANN")) install.packages("RANN")
library("RANN")

# Creation of x vector of longitude coordinates, y vector of latitude coordinates and finaly merge them on a position data frame.
x <- c(1,2,3,4,7.5,8,8.5,9,10,13,13.1,13.2,13.3,14,15,30)
y <- c(1,2,3,4,7.5,8,8.5,9,10,13,13.1,13.2,13.3,14,15,30)
pos <- data.frame(x,y)

# Creation of test data frame with 0 for negative cases and 1 for positive clases for each position.
test <- c(0,1,1,0,1,0,1,1,0,0,0,0,1,0,1,1)

# Creation of catalogue for this positions, linking distance 2 and default values.
cat <- catalogue(pos, test, 2)

scatter_pval(pos, cat$cluster_id, (test == 1), cat$epifriends_catalogue)

}
\author{
Eric Matamoros Morales based on earlier python code by Arnau Pujol.
}
