% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automate_link_d.R
\name{simplify_distributions}
\alias{simplify_distributions}
\title{Merge quantiles based on proximity of distances given its distribution.}
\usage{
simplify_distributions(dt, dist_prop)
}
\arguments{
\item{dt}{data.table with quantiles and minimum distances.}

\item{dist_prop}{Distance threshold to merge two quantiles due to proximity.}
}
\value{
data.table with simplified quantiles based on proximity.
}
\description{
Merge quantiles based on proximity of distances given its distribution.
}
\examples{

# Creation of x vector of longitude coordinates, y vector of latitude coordinates 
# and finaly merge them on a position data frame.
quantiles <- c(0, 0.125, 0.250, 0.5)
y <- c(1,2,3,4,7.5,8,8.5,9,10,13,13.1,13.2,13.3,14,15,30)
test <-c(0,1,1,0,1,0,1,1,0,0,0,0,1,0,1,1)

dt <- data.table(quantiles = quantiles, distances = c(0.01, 0.02, 0.03, 0.04))

# Simplify distributions based on proximity
distributions_new <- simplify_distributions(dt, 0.1)
}
\author{
Eric Matamoros.
}
