% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogue.R
\name{simulate_trial}
\alias{simulate_trial}
\title{Perform binomial simulations with different independent probabilities of success.}
\usage{
simulate_trial(n, size, probs)
}
\arguments{
\item{n}{Number of observations.}

\item{size}{Number of trials (zero or more).}

\item{probs}{Vector of probabilities for each simulation to be performed.}
}
\value{
Ratio between the sum of each n and the total number of trials.
}
\description{
Perform binomial simulations with different independent probabilities of success.
}
\examples{
# Creation of x vector of longitude coordinates, y vector of latitude coordinates 
# and finally merge them on a position data frame.
n <- 10000
size <- 1
probs <- c(0.3, 0.4)

# Do simulations
trials <- simulate_trial(n, size, probs)

}
\author{
Eric Matamoros.
}
