% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{size_histogram}
\alias{size_histogram}
\title{This method shows a histogram of the size (total number of cases) in foci for foci with p>0.05 and with p<0.05.}
\usage{
size_histogram(catalogue)
}
\arguments{
\item{catalogue}{Catalogue of the epifriends clusters and their main characteristics as output of the catalogue() function.}
}
\value{
Histogram of number of foci per total number of cases with p<0.05 in red and p>0.05 in blue.
}
\description{
This method shows a histogram of the size (total number of cases) in foci for foci with p>0.05 and with p<0.05.
}
\examples{
# Required packages
if(!require("RANN")) install.packages("RANN")
library("RANN")

# Creation of x vector of longitude coordinates, y vector of latitude coordinates and finaly merge them on a position data frame.
x <- c(1,2,3,4,7.5,8,8.5,9,10,13,13.1,13.2,13.3,14,15,30)
y <- c(1,2,3,4,7.5,8,8.5,9,10,13,13.1,13.2,13.3,14,15,30)
pos <- data.frame(x,y)

# Creation of test data frame with 0 for negative cases and 1 for positive clases for each position.
test <- data.frame(c(0,1,1,0,1,0,1,1,0,0,0,0,1,0,1,1))

# Creation of catalogue for this positions, linking distance 2 and default values.
cat <- catalogue(pos, test, 2)

size_histogram(cat)

}
\author{
Eric Matamoros Morales based on earlier python code by Arnau Pujol.
}
